+++ src/shared/meson.build
@@ -10,6 +10,10 @@ schedules_c = files([
   'schedules.c',
   ])
 
+pipes_c = files([
+	'pipes.c',
+	])
+
 if selinux_dep.found()
   selinux_c = files([
     'selinux.c',
